Removing Workflow
The reason for the two different calculations is because storage costs are based on an hourly aggregate of use. Costs are based on size and how long it's stored, and not just size alone.

For example, If I upload 3 GB and delete it after 3 days, I'll have accumulated 72 hours' worth of usage costs for that 3 GB.

Read more about how storage is calculated here:

https://docs.github.com/en/billing/managing-billing-for-github-actions/about-billing-for-github-actions#calculating-minute-and-storage-spending

Shared storage consists of Actions artifacts and Packages usage. You can download a detailed usage report from the billing page:

https://github.com/settings/billing

Here are the steps:

https://docs.github.com/en/github/setting-up-and-managing-billing-and-payments-on-github/viewing-your-github-actions-usage

In this case, it looks like the usage is coming from artifacts. Artifacts from the Meetly repository appear to be consuming the majority of the storage.

By default, artifacts are deleted on a rolling 90-day basis. That can be adjusted down to as little as 1-day:

https://docs.github.com/en/actions/guides/storing-workflow-data-as-artifacts#configuring-a-custom-artifact-retention-period

You can also delete artifacts via API:

https://developer.github.com/v3/actions/artifacts/#delete-an-artifact

Your monthly Actions minutes and bandwidth usage resets each month on the billing date. Existing storage will carry over to the next month if the objects remain stored in the account. If you find that you've exhausted your monthly included usage and you'd like to continue to use GitHub Actions and Packages between now and your next billing date, you can increase the spending limit:

https://help.github.com/en/github/setting-up-and-managing-billing-and-payments-on-github/managing-your-spending-limit-for-github-actions